requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
asyncio==3.4.3
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2021.10.8
chardet==4.0.0
click==8.1.2
dash==2.3.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
dill==0.3.4
diskcache==5.4.0
et-xmlfile==1.1.0
fastapi==0.76.0
Flask==2.1.1
Flask-Compress==1.11
fyers-apiv2==2.0.5
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
multiprocess==0.70.12.2
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
plotly==5.7.0
pony==0.7.16
psutil==5.9.0
psycopg2-binary==2.9.3
pydantic==1.9.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.1
requests==2.25.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
starlette==0.18.0
tenacity==8.0.1
tornado==6.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websocket-client==1.2.1
websockets==10.3
Werkzeug==2.1.1
wrapt==1.14.1
zipp==3.8.0
